Console avatar opens the same account menu as the app
The rail's AccountMenu moves to its own file with a placement prop;
the console header's inert avatar span becomes that same control, and
the header's standalone Sign out button folds into it.
