+++ frontend/src/components/workspace/rail.tsx
@@ -4,6 +4,7 @@ import { Fragment, useEffect, useRef, useState } from "react";
 import Link from "next/link";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import { Bot, FolderTree, MessagesSquare, GraduationCap, Home, Wrench, Settings } from "lucide-react";
+import AccountMenu from "@/components/workspace/account-menu";
 import { cn } from "@/lib/utils";
 import { useEscapeKey } from "@/hooks/useEscapeKey";
 import { useWorkspace, type RailSection } from "@/lib/workspace-store";
@@ -53,46 +54,6 @@ function RailButton({
       <Icon className="h-[18px] w-[18px]" />
       <span className="text-[10px] font-medium leading-none">{item.label}</span>
     </button>
-  );
-}
-
-/** Bottom-left account control — avatar opens a small menu (settings + sign out).
- *  This is the single home for account actions (removed from the top bar). */
-function AccountMenu({ user, onLogout }: { user: User; onLogout: () => void }) {
-  const [open, setOpen] = useState(false);
-  const ref = useRef<HTMLDivElement>(null);
-  useEscapeKey(open, () => setOpen(false));
-  useEffect(() => {
-    if (!open) return;
-    const onDown = (e: MouseEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node)) setOpen(false);
-    };
-    document.addEventListener("mousedown", onDown);
-    return () => document.removeEventListener("mousedown", onDown);
-  }, [open]);
-  return (
-    <div ref={ref} className="relative">
-      <button
-        onClick={() => setOpen((o) => !o)}
-        title={user.email ?? user.name}
-        className="flex h-8 w-8 items-center justify-center rounded-full bg-brand-500 text-[12px] font-semibold text-white hover:ring-2 hover:ring-brand-200"
-      >
-        {user.display_name[0].toUpperCase()}
-      </button>
-      {open && (
-        <div role="menu" className="absolute bottom-0 left-full z-40 ml-2 w-56 overflow-hidden rounded-md border border-border bg-surface py-1 text-[13px] shadow-lg">
-          <div className="border-b border-border px-3 py-1.5 text-[11px] text-muted-foreground">
-            Signed in as <span className="break-all text-foreground">{user.email ?? user.name}</span>
-          </div>
-          <Link href="/settings" onClick={() => setOpen(false)} className="block px-3 py-1.5 text-foreground hover:bg-raised">
-            Account settings
-          </Link>
-          <button onClick={() => { setOpen(false); onLogout(); }} className="block w-full px-3 py-1.5 text-left text-foreground hover:bg-raised">
-            Sign out
-          </button>
-        </div>
-      )}
-    </div>
   );
 }
